refactor: replace TableData/CourseData cascade stubs with real imports (Phase 15g)
Replace CourseData, TrophyData, CourseDataConstraint, TableData, TableFolder,
TableAccessor, TableDataAccessor, DifficultyTableAccessor, and CourseDataAccessor
stubs in beatoraja-select with real imports from beatoraja-types and beatoraja-core.

- Add missing getter/setter methods to real CourseData and TrophyData
- Add Send + Sync bounds to TableAccessor trait
- Update BMSSearchAccessor to implement real TableAccessor trait
- Update table_bar.rs to use get_url_opt() for Option<&str>
- Update grade_bar.rs TrophyData arithmetic from f64 to f32

+++ lr2oraja-rust/crates/beatoraja-select/src/bar/grade_bar.rs
@@ -82,9 +82,9 @@ impl GradeBar {
 
     fn qualified(score: &ScoreData, trophy: &TrophyData) -> bool {
         score.get_notes() != 0
-            && trophy.get_missrate() >= score.get_minbp() as f64 * 100.0 / score.get_notes() as f64
+            && trophy.get_missrate() >= score.get_minbp() as f32 * 100.0 / score.get_notes() as f32
             && trophy.get_scorerate()
-                <= score.get_exscore() as f64 * 100.0 / (score.get_notes() as f64 * 2.0)
+                <= score.get_exscore() as f32 * 100.0 / (score.get_notes() as f32 * 2.0)
     }
 
     pub fn get_lamp(&self, _is_player: bool) -> i32 {
